# Manager - improve performance apply errata bsc 1106430