--- a/java/code/src/com/redhat/rhn/manager/action/ActionChainManager.java
+++ b/java/code/src/com/redhat/rhn/manager/action/ActionChainManager.java
@@ -622,7 +622,7 @@ public class ActionChainManager {
                 taskomaticApi.scheduleActionExecution(action);
             }
             if (ActionFactory.TYPE_PACKAGES_UPDATE.equals(type)) {
-                MinionActionManager.scheduleStagingJobsForMinions(action, user);
+                MinionActionManager.scheduleStagingJobsForMinions(Collections.singletonList(action), user);
             }
         }
         else {
@@ -636,7 +636,7 @@ public class ActionChainManager {
                     nextSortOrder);
                 result.add(action);
                 if (ActionFactory.TYPE_PACKAGES_UPDATE.equals(type)) {
-                    MinionActionManager.scheduleStagingJobsForMinions(action, user);
+                    MinionActionManager.scheduleStagingJobsForMinions(Collections.singletonList(action), user);
                 }
             }
         }