… java/code/src/com/redhat/rhn/common/db/datasource/xml/Errata_queries.xml
… java/code/src/com/redhat/rhn/common/db/datasource/xml/ssm_operation_queries.xml
… java/code/src/com/redhat/rhn/domain/errata/impl/PublishedErrata.hbm.xml
… java/code/src/com/redhat/rhn/domain/errata/impl/UnpublishedErrata.hbm.xml
… java/code/src/com/redhat/rhn/domain/server/Server_legacyUser.hbm.xml
… java/code/src/com/redhat/rhn/manager/action/test/MinionActionManagerTest.java
… java/code/src/com/redhat/rhn/manager/errata/test/ErrataManagerTest.java
… java/code/src/com/redhat/rhn/taskomatic/TaskoXmlRpcHandler.java
… java/spacewalk-java.changes