## What does this PR change?

Improves the performance when scheduling a patching action

The following tests were executed for performance analysis before and after applying the proposed changes:

All the tests were executed with the flag "enabled content staging" in TRUE, and all the tests were executed over a system set of 10.5k minions, which is the size of the systems set, NOT the size of the affected systems.

 Test                                                | Total time BEFORE changes     | Total time AFTER changes     |
|-----------------------------------------------------|----------|----------|
| Patching: 1 patch, ~1.5k affected minions, 1 scheduled action | ~109 minutes | ~1 minute |
| Patching: 500 patches, ~7k affected minions, 712 scheduled actions | ~117 minutes | ~5 minutes |
| Package upgrading: 1 package, ~1.5k affected minions, 1 scheduled actions | ~13 minutes | ~35 seconds |
| Package upgrading: 500 packages, ~3k affected minions, 58 scheduled actions  | ~16 minutes | ~1 minute 20 seconds |

## Documentation
- No documentation needed: Bug fix

- [x] **DONE**

## Test coverage
- No tests: unit test exists already

- [x] **DONE**

## Links

Fixes https://github.com/SUSE/spacewalk/issues/5640
Tracks https://github.com/SUSE/spacewalk/pull/5910 and https://github.com/SUSE/spacewalk/pull/6080

- [x] **DONE**